--- src/main/java/DukeException.java
@@ -1,7 +1,0 @@
-public class DukeException extends Exception{
-    private String description;
-
-    DukeException(String description) {
-        this.description = description;
-    }
-}
