--- src/main/java/Task.java
@@ -1,25 +1,0 @@
-import java.util.Arrays;
-
-public class Task {
-    protected String description;
-    protected boolean isDone;
-
-    public Task(String description) {
-        this.description = description;
-        this.isDone = false;
-    }
-
-    public String getStatusIcon() {
-        return (isDone ? "X" : " "); // mark done task with X
-    }
-
-    public void markAsDone() {
-        this.isDone = true;
-    }
-
-    @Override
-    public String toString(){
-        return String.format("[%s] %s", this.getStatusIcon() , this.description);
-    }
-}
-
