+++ src/main/java/duke/DukeException.java
@@ -1,0 +1,9 @@
+package duke;
+
+public class DukeException extends Exception {
+    private final String description;
+
+    DukeException(String description) {
+        this.description = description;
+    }
+}
