+++ src/main/java/duke/Task.java
@@ -1,0 +1,33 @@
+package duke;
+
+/**
+ * Task class used by the duke class
+ * Stores a description of task and whether it is done
+ */
+public class Task {
+    protected String description;
+    protected boolean isDone;
+
+    public Task(String description) {
+        this.description = description;
+        this.isDone = false;
+    }
+
+    /**
+     * Given the isDone status return appropriate icon
+     *
+     * @return String "X" if task is done " " if not done
+     */
+    public String getStatusIcon() {
+        return (isDone ? "X" : " "); // mark done task with X
+    }
+
+    public void markAsDone() {
+        this.isDone = true;
+    }
+
+    @Override
+    public String toString() {
+        return String.format("[%s] %s", this.getStatusIcon(), this.description);
+    }
+}
